+++ mochawesome-report/node02_dkg.html
@@ -1,0 +1,2 @@
+<!doctype html>
+<html lang="en"><head><meta charSet="utf-8"/><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width, initial-scale=1"/><title>Mochawesome Report</title><link rel="stylesheet" href="assets/app.css"/></head><body data-raw="{&quot;stats&quot;:{&quot;suites&quot;:1,&quot;tests&quot;:1,&quot;passes&quot;:1,&quot;pending&quot;:0,&quot;failures&quot;:0,&quot;start&quot;:&quot;2025-04-22T14:14:43.559Z&quot;,&quot;end&quot;:&quot;2025-04-22T14:15:15.834Z&quot;,&quot;duration&quot;:32275,&quot;testsRegistered&quot;:1,&quot;passPercent&quot;:100,&quot;pendingPercent&quot;:0,&quot;other&quot;:0,&quot;hasOther&quot;:false,&quot;skipped&quot;:0,&quot;hasSkipped&quot;:false},&quot;results&quot;:[{&quot;uuid&quot;:&quot;61a0d39b-8233-4d45-bba4-5e1f837783bd&quot;,&quot;title&quot;:&quot;&quot;,&quot;fullFile&quot;:&quot;/Users/bojancolic/dkg.js/tests/publish-ka-testnet-node02.spec.js&quot;,&quot;file&quot;:&quot;/tests/publish-ka-testnet-node02.spec.js&quot;,&quot;beforeHooks&quot;:[],&quot;afterHooks&quot;:[],&quot;tests&quot;:[],&quot;suites&quot;:[{&quot;uuid&quot;:&quot;eba4d46d-9d9b-4139-a357-ee51a2fda8a9&quot;,&quot;title&quot;:&quot;DKG Asset Lifecycle on Testnet&quot;,&quot;fullFile&quot;:&quot;/Users/bojancolic/dkg.js/tests/publish-ka-testnet-node02.spec.js&quot;,&quot;file&quot;:&quot;/tests/publish-ka-testnet-node02.spec.js&quot;,&quot;beforeHooks&quot;:[],&quot;afterHooks&quot;:[],&quot;tests&quot;:[{&quot;title&quot;:&quot;should publish and query a Knowledge Asset successfully&quot;,&quot;fullTitle&quot;:&quot;DKG Asset Lifecycle on Testnet should publish and query a Knowledge Asset successfully&quot;,&quot;timedOut&quot;:false,&quot;duration&quot;:32272,&quot;state&quot;:&quot;passed&quot;,&quot;speed&quot;:&quot;slow&quot;,&quot;pass&quot;:true,&quot;fail&quot;:false,&quot;pending&quot;:false,&quot;code&quot;:&quot;const content = {\n  public: {\n    &#x27;@context&#x27;: &#x27;https://www.schema.org&#x27;,\n    &#x27;@id&#x27;: &#x27;urn:first-ka:info:hello-dkg&#x27;,\n    &#x27;@type&#x27;: &#x27;CreativeWork&#x27;,\n    name: &#x27;Hello DKG&#x27;,\n    description: &#x27;My first Knowledge Asset on the Decentralized Knowledge Graph!&#x27;,\n  },\n};\nconst create_result = await DkgClient.asset.create(content, {\n  epochsNum: 2,\n  minimumNumberOfFinalizationConfirmations: 3,\n  minimumNumberOfNodeReplications: 1,\n});\nif (!create_result) {\n  throw new Error(&#x27;Knowledge Asset was not published successfully&#x27;);\n}\n\n  console.log(JSON.stringify(create_result));\nconst queryResult = await DkgClient.graph.query(`\n  PREFIX schema: &lt;http://schema.org/&gt;\n  SELECT ?s ?name ?description\n  WHERE {\n    ?s schema:name ?name ;\n       schema:description ?description .\n  }\n`, &#x27;SELECT&#x27;);\nassert.ok(queryResult?.data?.length &gt; 0, &#x27;Query returned no results&#x27;);\nconsole.log(&#x27;Query results:&#x27;, queryResult);&quot;,&quot;err&quot;:{},&quot;uuid&quot;:&quot;522a0d5b-8c70-4672-8aef-3a466a7cde57&quot;,&quot;parentUUID&quot;:&quot;eba4d46d-9d9b-4139-a357-ee51a2fda8a9&quot;,&quot;isHook&quot;:false,&quot;skipped&quot;:false}],&quot;suites&quot;:[],&quot;passes&quot;:[&quot;522a0d5b-8c70-4672-8aef-3a466a7cde57&quot;],&quot;failures&quot;:[],&quot;pending&quot;:[],&quot;skipped&quot;:[],&quot;duration&quot;:32272,&quot;root&quot;:false,&quot;rootEmpty&quot;:false,&quot;_timeout&quot;:360000}],&quot;passes&quot;:[],&quot;failures&quot;:[],&quot;pending&quot;:[],&quot;skipped&quot;:[],&quot;duration&quot;:0,&quot;root&quot;:true,&quot;rootEmpty&quot;:true,&quot;_timeout&quot;:2000}],&quot;meta&quot;:{&quot;mocha&quot;:{&quot;version&quot;:&quot;11.1.0&quot;},&quot;mochawesome&quot;:{&quot;options&quot;:{&quot;quiet&quot;:false,&quot;reportFilename&quot;:&quot;node02_dkg&quot;,&quot;saveHtml&quot;:true,&quot;saveJson&quot;:true,&quot;consoleReporter&quot;:&quot;spec&quot;,&quot;useInlineDiffs&quot;:false,&quot;code&quot;:true},&quot;version&quot;:&quot;7.1.3&quot;},&quot;marge&quot;:{&quot;options&quot;:{&quot;reportFilename&quot;:&quot;node02_dkg&quot;},&quot;version&quot;:&quot;6.2.0&quot;}}}" data-config="{&quot;reportFilename&quot;:&quot;node02_dkg&quot;,&quot;reportDir&quot;:&quot;mochawesome-report&quot;,&quot;reportTitle&quot;:&quot;dkg.js&quot;,&quot;reportPageTitle&quot;:&quot;Mochawesome Report&quot;,&quot;inline&quot;:false,&quot;inlineAssets&quot;:false,&quot;cdn&quot;:false,&quot;charts&quot;:false,&quot;enableCharts&quot;:false,&quot;code&quot;:true,&quot;enableCode&quot;:true,&quot;autoOpen&quot;:false,&quot;overwrite&quot;:true,&quot;timestamp&quot;:false,&quot;ts&quot;:false,&quot;showPassed&quot;:true,&quot;showFailed&quot;:true,&quot;showPending&quot;:true,&quot;showSkipped&quot;:false,&quot;showHooks&quot;:&quot;failed&quot;,&quot;saveJson&quot;:true,&quot;saveHtml&quot;:true,&quot;dev&quot;:false,&quot;assetsDir&quot;:&quot;mochawesome-report/assets&quot;,&quot;jsonFile&quot;:&quot;/Users/bojancolic/dkg.js/mochawesome-report/node02_dkg.json&quot;,&quot;htmlFile&quot;:&quot;/Users/bojancolic/dkg.js/mochawesome-report/node02_dkg.html&quot;}"><div id="report"></div><script src="assets/app.js"></script></body></html>
