+++ mochawesome-report/node02_dkg.json
@@ -1,0 +1,102 @@
+{
+  "stats": {
+    "suites": 1,
+    "tests": 1,
+    "passes": 1,
+    "pending": 0,
+    "failures": 0,
+    "start": "2025-04-22T14:14:43.559Z",
+    "end": "2025-04-22T14:15:15.834Z",
+    "duration": 32275,
+    "testsRegistered": 1,
+    "passPercent": 100,
+    "pendingPercent": 0,
+    "other": 0,
+    "hasOther": false,
+    "skipped": 0,
+    "hasSkipped": false
+  },
+  "results": [
+    {
+      "uuid": "61a0d39b-8233-4d45-bba4-5e1f837783bd",
+      "title": "",
+      "fullFile": "/Users/bojancolic/dkg.js/tests/publish-ka-testnet-node02.spec.js",
+      "file": "/tests/publish-ka-testnet-node02.spec.js",
+      "beforeHooks": [],
+      "afterHooks": [],
+      "tests": [],
+      "suites": [
+        {
+          "uuid": "eba4d46d-9d9b-4139-a357-ee51a2fda8a9",
+          "title": "DKG Asset Lifecycle on Testnet",
+          "fullFile": "/Users/bojancolic/dkg.js/tests/publish-ka-testnet-node02.spec.js",
+          "file": "/tests/publish-ka-testnet-node02.spec.js",
+          "beforeHooks": [],
+          "afterHooks": [],
+          "tests": [
+            {
+              "title": "should publish and query a Knowledge Asset successfully",
+              "fullTitle": "DKG Asset Lifecycle on Testnet should publish and query a Knowledge Asset successfully",
+              "timedOut": false,
+              "duration": 32272,
+              "state": "passed",
+              "speed": "slow",
+              "pass": true,
+              "fail": false,
+              "pending": false,
+              "context": null,
+              "code": "const content = {\n  public: {\n    '@context': 'https://www.schema.org',\n    '@id': 'urn:first-ka:info:hello-dkg',\n    '@type': 'CreativeWork',\n    name: 'Hello DKG',\n    description: 'My first Knowledge Asset on the Decentralized Knowledge Graph!',\n  },\n};\nconst create_result = await DkgClient.asset.create(content, {\n  epochsNum: 2,\n  minimumNumberOfFinalizationConfirmations: 3,\n  minimumNumberOfNodeReplications: 1,\n});\nif (!create_result) {\n  throw new Error('Knowledge Asset was not published successfully');\n}\n\n  console.log(JSON.stringify(create_result));\nconst queryResult = await DkgClient.graph.query(`\n  PREFIX schema: <http://schema.org/>\n  SELECT ?s ?name ?description\n  WHERE {\n    ?s schema:name ?name ;\n       schema:description ?description .\n  }\n`, 'SELECT');\nassert.ok(queryResult?.data?.length > 0, 'Query returned no results');\nconsole.log('Query results:', queryResult);",
+              "err": {},
+              "uuid": "522a0d5b-8c70-4672-8aef-3a466a7cde57",
+              "parentUUID": "eba4d46d-9d9b-4139-a357-ee51a2fda8a9",
+              "isHook": false,
+              "skipped": false
+            }
+          ],
+          "suites": [],
+          "passes": [
+            "522a0d5b-8c70-4672-8aef-3a466a7cde57"
+          ],
+          "failures": [],
+          "pending": [],
+          "skipped": [],
+          "duration": 32272,
+          "root": false,
+          "rootEmpty": false,
+          "_timeout": 360000
+        }
+      ],
+      "passes": [],
+      "failures": [],
+      "pending": [],
+      "skipped": [],
+      "duration": 0,
+      "root": true,
+      "rootEmpty": true,
+      "_timeout": 2000
+    }
+  ],
+  "meta": {
+    "mocha": {
+      "version": "11.1.0"
+    },
+    "mochawesome": {
+      "options": {
+        "quiet": false,
+        "reportFilename": "node02_dkg",
+        "saveHtml": true,
+        "saveJson": true,
+        "consoleReporter": "spec",
+        "useInlineDiffs": false,
+        "code": true
+      },
+      "version": "7.1.3"
+    },
+    "marge": {
+      "options": {
+        "reportFilename": "node02_dkg"
+      },
+      "version": "6.2.0"
+    }
+  }
+}
