+++ package.json
@@ -15,7 +15,10 @@
         "lint": "eslint .",
         "cjs-compat": "npx rollup index.js --file index.cjs --format cjs",
         "postinstall": "npm run cjs-compat",
-        "test:publish:query:ka:testnet": "node publish-ka-testnet-node02.js && node publish-ka-testnet-node03.js",
+        "test:dkg:node02": "mocha tests/publish-ka-testnet-node02.spec.js --reporter mochawesome --reporter-options reportFilename=node02_dkg",
+        "test:dkg:node03": "mocha tests/publish-ka-testnet-node03.spec.js --reporter mochawesome --reporter-options reportFilename=node03_dkg",
+        "generate:testnet:node02:report": "node report-testnet-node02-config.js",
+        "generate:testnet:node03:report": "node report-testnet-node03-config.js",
         "test:publish:query:ka:mainnet": "node publish-ka-mainnet.js"
     },
     "repository": {
@@ -41,7 +44,6 @@
         "assertion-tools": "8.0.2",
         "axios": "^0.27.2",
         "dkg-evm-module": "8.0.4",
-        "dotenv": "^16.4.7",
         "ethers": "^6.1.0",
         "jsonld": "^8.1.0",
         "web3": "^1.7.3"
@@ -50,16 +52,20 @@
         "assert": "^2.0.0",
         "buffer": "^6.0.3",
         "crypto-browserify": "^3.12.0",
+        "dotenv": "^16.5.0",
         "eslint": "^8.23.0",
         "eslint-config-airbnb": "^19.0.4",
         "eslint-config-prettier": "^8.5.0",
         "https-browserify": "^1.0.0",
+        "mocha": "^11.1.0",
+        "mochawesome": "^7.1.3",
         "os-browserify": "^0.3.0",
         "prettier": "^2.7.1",
         "rollup": "^4.28.1",
         "stream-browserify": "^3.0.0",
         "stream-http": "^3.2.0",
         "terser-webpack-plugin": "^5.3.6",
+        "test-results-reporter": "^1.1.6",
         "url": "^0.11.1",
         "webpack": "^5.75.0",
         "webpack-cli": "^4.9.2"
